Add audioAutoplay configuration option
Sometimes it is beneficial to start audio narration immediately,
without waiting for the user to click the play button in the audio
player. Add a new configuration option, audioAutoplay, that will
start playing audio immediately upon the Reveal.ready event.

When set, this will also autoplay any audio on a slide that is being
jumped to directly, either from overview mode, or via a direct URL.

Default to false, retaining the default behavior of auto-playing only
when advancing to a new slide, with the user having to click the play
button at least once on a preceding slide.

Breaking what was previously a single if block into a nested one is
necessary to not have the audio loop indefinitely on the final slide,
when audioAutoplay is set.

Fixes rajgoel/reveal.js#10.