Notes: automatically adjust total time if presentation starts late (or early)
Sometimes, the beginning of a presentation is delayed for technical or
other reasons. Frequently, under such circumstances, speakers are
still expected to finish their talk on time.

This patch modifies the behavior of the totalTime configuration option
in the following way:

* There is a new configuration option, scheduledStart. This is meant
  to be set to the time the presentation is scheduled to start.

* There is also scheduledEnd, which is a convenience option: if you
  specify both scheduledEnd and scheduledStart, then you do not need
  to set totalTime; it is calculated automatically.

* If scheduledStart is set, then the local system time is evaluated at
  the start of the presentation, and every time you click the reset
  button on the clock. When this happens, totalTime is automatically
  adjusted. Thus, if you start your presentation a minute or two late,
  then the allotted time for each slide decreases by a few
  seconds. This enables you to make up the time lost at the beginning
  bit by bit, rather than being short on time at the end.

scheduledStart must be given in local time, or more precisely, in the
timezone you expect the presentation computer to be set to while
presenting. That is to say, if your laptop always uses UTC or your
home timezone, and you never change it while traveling into other
timezones, then use your laptop timezone.

Adjusting the slide timing in this way obviously tends to be useless
during rehearsal. You would thus typically rehearse your talk with
just totalTime set, and then add the scheduledStart configuration
option immediately before delivering the presentation.